pyproject.toml
[project]
name = "cryptosignalagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.3.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langsmith>=0.7.9",
    "llama-index>=0.14.15",
    "llama-index-vector-stores-milvus>=1.0.0",
    "mem0ai>=1.0.4",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pymilvus>=2.6.9",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.41.0",
    "zhipuai>=2.1.5",
]
